backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langgraph==0.2.16
langchain==0.2.16
langchain-groq==0.1.9
langchain-core==0.2.38
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
groq==0.4.1
